HW/assignment3/requirements.txt
box2d-py==2.3.5
certifi==2022.12.7
charset-normalizer==2.1.1
cloudpickle==2.2.0
cmaes==0.9.0
gym==0.26.2
gym-notices==0.0.8
idna==3.4
imageio==2.22.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
pandas==1.5.2
Pillow==9.3.0
pygame==2.1.0
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
swig==4.1.0
torch==1.13.0
torchvision==0.14.0
typing_extensions==4.4.0
urllib3==1.26.13
